公司信息显示

+++ ecltproject/src/com/eclt/dao/EcCompanyDao.java
@@ -6,6 +6,6 @@ import com.bstek.dorado.hibernate.HibernateDao;
 import com.eclt.entity.EcCompany;
 
 @Repository
-public class EcCompanyDao extends HibernateDao<EcCompany, Long> {
+public class EcCompanyDao extends HibernateDao<EcCompany, Integer> {
 
 }
